Fix install script, use .dotfiles folder

+++ makesymlinks.sh
@@ -6,9 +6,9 @@
 
 ########## Variables
 
-dir=~/dotfiles             # dotfiles directory
-olddir=~/dotfiles_old      # old dotfiles backup directory
-files="vimrc vim zshrc"    # list of files/folders to symlink in homedir
+dir=~/.dotfiles             # dotfiles directory
+olddir=~/.dotfiles_old      # old dotfiles backup directory
+files="vimrc zshrc"    # list of files/folders to symlink in homedir
 
 ##########
 
@@ -25,8 +25,10 @@ echo "...done"
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks
 for file in $files; do
     echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
+    mv ~/.$file $olddir/
     echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [ ! -e $file ]; then
+        ln -s $dir/$file ~/.$file
+    fi
 done
